ci: frontend-desing-refactoring
