+++ cli_application/src/components/pages/UserAccountsListPage.jsx
@@ -1,0 +1,13 @@
+import React from 'react'
+
+const UserAccountsListPage = () =>{
+    return(
+        <div>
+            <h1>User Accounts</h1>
+        </div>
+    )
+}
+
+
+export default UserAccountsListPage;
+
