+++ cli_application/src/components/panel/adminpanel.css
@@ -1,16 +1,43 @@
 .item-profile{
-    background-color: #f4eeff;
-    border-right: 2px solid #5351d6;
+    background-color: #ffffff;
+    border-right: 2px solid #5351c4;
     border-top: none;
     border-bottom: none;
     border-left: none;
+    
 }
 
-
-.sub-item-option{
-    background-color: #e9f0ff;
-    border-right: 2px solid #fab433;
-    border-top: none;
-    border-bottom: none;
-    border-left: none;
+.elemento-padre {
+    position: relative;
+    display: block;
+    padding: 0.75rem 1rem;
+    background-color: #fff;
+    border-top-width: 1px;
+    border-right: 4px solid #3357fa;
+    margin-bottom: 0.6rem;
+    margin-right: 0.4rem;
+  
 }
+
+.elemento-padre:hover{
+    text-decoration: none;
+}
+
+.elemento-hijo {
+    position: relative;
+    display: block;
+    padding: 0.75rem 1rem;
+    background-color: #fff;
+    border-top-width: 1px;
+    border-right: 4px solid #fab433;
+    margin-bottom: 0.4rem;
+    margin-right: 0.4rem;
+    margin-left: 0.4rem;
+    margin-top: 0.4rem;
+}
+
+.elemento-hijo:hover{
+    text-decoration: none;
+    z-index: 1;
+}
+
