+++ cli_application/src/components/user/Profile/UserProfile.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+
+const UserProfile  = () =>{
+    return(
+        <h1 className="border py-3 px-3 shadow">Perfil del usuario</h1>
+    )
+}
+
+export default UserProfile;
